requirements.txt
altair==4.2.0
appnope==0.1.3
argcomplete==2.0.0
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
beautifulsoup4==4.11.1
bertopic==0.13.0
blinker==1.5
bqplot==0.12.36
branca==0.6.0
cachetools==5.2.1
charset-normalizer==3.0.1
click==8.1.3
click-plugins==1.1.1
cligj==0.7.2
colour==0.1.5
comm==0.1.2
commonmark==0.9.1
contourpy==1.0.7
cycler==0.11.0
Cython==0.29.33
debugpy==1.6.5
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
Fiona==1.8.22
folium==0.14.0
fonttools==4.38.0
gdown==4.6.0
geojson==2.5.0
geopandas==0.12.2
gitdb==4.0.10
GitPython==3.1.30
hdbscan==0.8.29
huggingface-hub==0.11.1
idna==3.4
importlib-metadata==6.0.0
ipyevents==2.0.1
ipyfilechooser==0.6.0
ipykernel==6.20.2
ipyleaflet==0.17.2
ipython==8.8.0
ipytree==0.2.2
ipywidgets==8.0.4
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.3
jupyter_client==7.4.9
jupyter_core==5.1.3
jupyterlab-widgets==3.0.5
kiwisolver==1.4.4
leafmap==0.15.0
llvmlite==0.39.1
MarkupSafe==2.1.1
matplotlib==3.6.3
matplotlib-inline==0.1.6
munch==2.5.0
nest-asyncio==1.5.6
nltk==3.8.1
numba==0.56.4
numpy==1.23.5
packaging==23.0
pandas==1.5.2
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pipx==1.0.0
platformdirs==2.6.2
plotly==5.12.0
prompt-toolkit==3.0.36
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pyaes==1.6.1
pyarrow==10.0.1
pyasn1==0.4.8
pydeck==0.8.0
Pygments==2.14.0
Pympler==1.0.1
pynndescent==0.5.8
pyogrio==0.5.0
pyparsing==3.0.9
pyproj==3.4.1
pyrsistent==0.19.3
pyshp==2.3.1
PySocks==1.7.1
pystac==1.6.1
pystac-client==0.5.1
python-box==6.1.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.4.1
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
rich==13.1.0
rsa==4.9
scikit-learn==1.2.0
scipy==1.10.0
scooby==0.7.0
semver==2.13.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
shapely==2.0.0
six==1.16.0
smmap==5.0.0
soupsieve==2.3.2.post1
stack-data==0.6.2
streamlit==1.17.0
Telethon==1.26.1
tenacity==8.1.0
threadpoolctl==3.1.0
tokenizers==0.13.2
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.1
traittypes==0.2.1
transformers==4.25.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
umap-learn==0.5.3
urllib3==1.26.14
userpath==1.8.0
validators==0.20.0
wcwidth==0.2.6
whitebox==2.2.0
whiteboxgui==2.2.0
widgetsnbextension==4.0.5
xyzservices==2022.9.0
zipp==3.11.0
